refactor(e2e): split smoke.spec.ts into 7 feature-scoped files
肥大化していた e2e/smoke.spec.ts(292 行・29 テスト)を機能別に 7 ファイル
へ分割する。テスト内容は維持し、ファイル単位での保守性とテスト発見性を
向上させる。

- e2e/home.spec.ts — トップページ(7 tests)
- e2e/strategy-detail.spec.ts — 戦略詳細ページ(3 tests)
- e2e/glossary.spec.ts — 用語ツールチップ + 用語集(4 tests)
- e2e/columns.spec.ts — コラム表示(1 test)
- e2e/search-and-404.spec.ts — 検索 + 404(2 tests)
- e2e/navigation.spec.ts — ナビゲーション(8 tests)
- e2e/toc.spec.ts — 目次(4 tests)

CLAUDE.md のテスト件数記載(16 → 29)も更新。

体感変化のないリファクタなので changelog 更新は不要(memory rule 8)。
